Added certificate signature algorithm override for DTLS
Added support for overriding the x509 certificate signature algorithm via system property when use in DTLS with WebRTC. If the system property is not overridden, the "old" value of SHA1withRSA is used. Allowing this to be overridden provides better support for Chrome and Firefox. To use the override, add this Java option to your startup: -Dorg.jitsi.impl.transform.dtls.SIGNATURE_ALGORITHM=SHA256withRSA