Introduces a REDBlockIterator.
In order to reduce code duplication this commit introduces a reusable
REDBlockIterator. The REDBlockIterator also offers a static utility
method that can be used to get the primary block of a RED packet
without having to instantiate an iterator an loop through the RED
blocks.

This is particularly interesting when this iterator is used to process
RED streams from Chrome. Chrome uses a bizzare encapsulation where each
RED packet contains either a single VP8 packet or a single FEC packet.
The REDBlockIterator.getPrimaryBlock() utility takes advantage of this
knowledge and doesn't iterate through the RED blocks in order to get the
primary block.