Reads all available application data from DTLS.
Reads all available application data from the DTLS transport, as opposed
to just a single datagram. This works around an issue causing the SCTP
connection in jitsi-videobridge to be delayed by around 13 seconds
(until the SCTP client retransmits the COOKIE-ECHO part of the
handshake).
A race condition between the DTLS connect thread and the SCTP thread of
jitsi-videobridge would result in a packet left in DatagramTransportImpl's
queue and not immediately decoded and returned to SCTP. This packet
would only be read when the next DTLS packet was received (at which time
this new packet was queued and not processed).