Set format on each source filter pin
Some webcams (or virtual webcam software) add an output pin as soon
as the first pin gets connected to a renderer. Worse, the new pin is
added at index 0. And worst, changing the format on the stream config
of the graph doesn't apply the new format to the already connected
pin.

Fix this by:
- refactoring to have a method that sets the format by operating
  on a stream config
- setting the format on the graph's stream config as a default
- enumerate all output pins and change the format on every pin's
  stream config

And update the associated binaries.

Closes jitsi/jitsi#220