Always uplifts RTP timestamps upon rewriting.
This commit attempts to fix the (now very rare) scrambling video problem
that we have been observing in our deployments. In particular, there's
no reason to enforce a threshold for the RTP timestamp uplifting. If
packets with old timestamps reach a receiver's jitter buffer they can
break the decoding process.