Add support for real OSGi in JNIUtils
A native library is normally loaded and bound to the ClassLoader of
the class calling System.loadLibray. If this task is delegated to a
utility method like in JNIUtils, and JNIUtils itself is in a different
ClassLoader than the caller, the native library won't be available to
the caller.

Add a method that passes the calling class and call the Runtime via
reflection with this passed class so that the native library ends up
in the ClassLoader of the class that calls JNIUtils.