[RemoteBitrateEstimator] Use the RTP extension abs-send-time to correctly calculate the inter-departure time.
Jitsi uses the RTP timestamp for the calculation of inter-departure
time.  This is not ideal because the RTP timestamp is recorded before
the pacer introduces artificial delay in the inter-arrival times.  We
fixed this by doing the same as Chrome does and use the RTP extension
abs-send-time to calculate deltas in the remote bitrate estimator.

With calls on meet.jit.si, the bitrate will unexpectedly drop.  This is
most noticeable with screenshare content in the following scenario:

1) Artificially high inter-group delay variation (d(i) in draft-ietf-rmcat-gcc-02)
because the departure times are recorded before the pacer queue.  Overuse is asserted.
2) If overuse occurs during a static screenshare scene, the bitrate
drops to 30kbps due to a ratcheting down to the low incoming bitrate.
3) High dynamic scene changes then cause the encoded bitrate to far
exceed the target bitrate.
4) The pacer limits the transmit rate to the target bitrate which can
cause the packets to be sent out over an interval of 5-10 seconds.

Test Plan:
Unit test covering the affected logic when we move to the 24-bit abs-send-time.
Manual tests of scenarios where the bitrate drop to 30kbps caused high
screenshare latency.

Reviewers: ed, ethan, hemanth

Subscribers: eng-team-list

Differential Revision: https://phab.fatline.io/D4188