[stats] Make incoming rate and BandwidthUsage stats per-client instead of per-stream.
Summary:
Incoming rate is supposed to be an estimate for the whole client.

BandwidthUsage is calculated per stream, but before we send the input
to the rate control, we aggregate and take the max(BandwidthUsage).
Might be easier to follow this stat if we just track one per the whole client.

Test Plan: View stats in MD.

Reviewers: ethan

Subscribers: eng-team-list

Differential Revision: https://phab.fatline.io/D4194