feat: Enables log correlation.
Inspired by the log4j NDC/MDC design this commits adds a couple of new
classes that can be used to provide contextual logging; the motivation
here is to allow/enable log corellation in the JVB server logs.
Contrary to the NDC/MDC paradigm, there can be different diagnostic
context providers in a LJ-based application. A diagnostic context
provider simply provides a diagnostic context in the NDC/MDC sense.
The diagnostic context is implemented in the `DiagnosticContext` class
and it has convenience methods to format time series points in the
influx DB line protocol format.