Fixes send-side bandwidth estimations and enables log mining. (#390)
* feat: Enables log correlation.

Inspired by the log4j NDC/MDC design this commits adds a couple of new
classes that can be used to provide contextual logging; the motivation
here is to allow/enable log corellation in the JVB server logs.
Contrary to the NDC/MDC paradigm, there can be different diagnostic
context providers in a LJ-based application. A diagnostic context
provider simply provides a diagnostic context in the NDC/MDC sense.
The diagnostic context is implemented in the `DiagnosticContext` class
and it has convenience methods to format time series points in the
influx DB line protocol format.

* fix: Outgoing RTCP TCC packet wire format.

This commit fixes an issue where the delta of the first acked packet is
0 due to the incorrect size of the packet chuncks buffer. This problem
occurs when the number of acked packets is a multiple of 7.

* style: Formatting.

* ref: Renames a method.

* ref: Renames a variable.

* logs: Adds diagnostic logging.

* fix: Unbreaks the initial probing phase when send-side cc is used.

The bridge is caching RTP packets while DTLS is being established. When
DTLS is established, the cached packets are decrypted and piggy-backed
onto some later media packet. Consequently, the order in which the SRTP
engine outputs packets is not the order in which the packets were
received.

Usually the cached packets are the first packets that are sent out by
chrome, and these first packets are part of chrome's probe clusters. It
turns out that the re-ordering breaks the initial probing phase of
webrtc/chrome, see:

https://bugs.chromium.org/p/webrtc/issues/detail?id=8720.

This PR makes sure that the transport-cc engine always runs before SRTP.
In order to do that, we split the transport-cc engine in two parts: the
ingress part which reads transport-cc sequence numbers before the RTP
packets reach the SRTP caching mechanism, and an egress part which adds
transport-cc sequence numbers before the RTP packets reach the SRTP
engine for encryption and authentication.

* Style, formatting and typos.

* logs: Uses the diangostic context for logging.

* style: Replaces requireNonNull with NotNull.

* logs: Adds a timestamp to some time series.

* logs: Makes the timestamp field of the diagnostic ctx final.

* logs: Adds the video media stream hashcode to the transport-cc engine diagnostic ctx.