AMR: Detect it within FFmpeg at Runtime.
OpenH264 is an extension to FFmpeg. The bundled FFmpeg for Microsoft Windows includes support for the video codec H.264. However, that bundled FFmpeg does not come with the audio codec AMR(-WB). The change for OpenH264 simply disabled AMR-WB in general, at compile time. Beside losing AMR-WB, that change was incomplete because AMR-WB is still added as a media format. For example, in Jitsi Desktop, AMR-WB is still selectable but none functional.

This change here re-enables AMR-WB but detects whether the used FFmpeg supports AMR-WB, at runtime. Consequently, when FFmpeg is missing or when the used FFmpeg does not support AMR-WB, it does not show up in Jitsi Desktop anymore. When FFmpeg is present, and when AMR-WB is supported, AMR-WB is listed and functional. On Linux, this allows the user to add/remove AMR-WB (or even FFmpeg) within the system without re-compiling Jitsi Desktop.