H.263+: Remove non-working commented setting.
The source code contains a line of code which is about the rate-control equation (rc_eq). The given one does not work with the default H.263+ encoder in Debian/Ubuntu and gives the error messages `Unknown function` and `Error parsing`. Because that line of code is commented/disabled anyway, it is simply removed.