FFmpeg: AVFilter: HFlip: Allow conversion between pixel formats.
This avoids the FFmpeg internal error message: The filters 'Parsed_buffer_0' and 'Parsed_hflip_1' do not have a common format and automatic conversion is disabled.

This happens, because the source (buffer) might use a pixel format which the filter (horizontal flip; hflip) is not able to use directly. Therefore, a conversion might be need which is enabled with the keyword scale. Because the sink (buffersink) shall have the same pixel format as the source, the same is required after the filter hflip.

closes jitsi/libjitsi#431
fixes jitsi/jitsi#540