H.264: Keep the default motion-estimation method.
With FFmpeg 2.7, the motion-estimation method (me_method) was moved from a public symbol to the private options of the underlying encoder interface for libx264. Since FFmpeg 4.0, the public symbol me_method does not exist anymore. Because Cisco OpenH264 ignored this anyway and the benefit for libx264 is unknown, the method is simply not changed anymore.