Fixes receiving remove videos.
The initial change was made in 912998b9, but was not taking in
consideration that all jitsi desktop video calls have a rtpTranslator,
and that change broke incoming video.
As this change is needed for jigasi, jigasi always sets useTranslator
for the audio and that setting is only used in that case.
fixes jitsi/jitsi#627
closes https://github.com/jitsi/libjitsi/pull/505