H.264: Set the motion-estimation method via private options. (#500)
* H.264: Set the motion-estimation method via private options.

With FFmpeg 2.7, the motion-estimation method (me_method) was moved from a public symbol to the private options of the underlying encoder interface for libx264. Since FFmpeg 4.0, the public symbol me_method does not exist anymore. Because the latest supported is FFmpeg is 2.8, simply use the private-option interface instead.

* H.264: Keep the default motion-estimation method.

With FFmpeg 2.7, the motion-estimation method (me_method) was moved from a public symbol to the private options of the underlying encoder interface for libx264. Since FFmpeg 4.0, the public symbol me_method does not exist anymore. Because Cisco OpenH264 ignored this anyway and the benefit for libx264 is unknown, the method is simply not changed anymore.