AMR: Encoder added a Null byte to Frame content. (#496)
* AMR: Encoder added a Null byte to Frame content.

Many AMR-WB implementations check the length of the frame, and reject frames with wrong lengths, even if only one frame is in the RTP packet. The solution is to use the right length which is dstLen - 1 = srcLen = the length has not changed.

fixes traud/asterisk-amr#17

* AMR: Encoder added a Null byte to Frame content.

Many AMR-WB implementations check the length of the frame, and reject frames with wrong lengths, even if only one frame is in the RTP packet. The solution is to use the correct length for the frame type which was actually created by the encoder.

fixes traud/asterisk-amr#17

* AMR: Encoder added a Null byte to Frame content.

Many AMR-WB implementations check the length of the frame, and reject frames with wrong lengths, even if only one frame is in the RTP packet. The solution is to use the correct length for the frame type which was actually created by the encoder.

fixes traud/asterisk-amr#17