fix: Skips ZRTTransformEngine when encryption is disabled. (#544)
* fix: Initialize the zrtp timeout thred only when needed.

In jigasi we saw some TimeoutProvider threads leaking ... where zrtp is not used. Seems there is a scenario in which the ZRTPTransformEngine is initialized but never stopped/cleaned. And zrtp is not used there.
Rarely happening, but still happening 20 times out of 2k.

* Revert "fix: Initialize the zrtp timeout thred only when needed."

This reverts commit 03a727b1911ecec61a909b1e008279c1daff22f6.

* fix: Skips ZRTTransformEngine when encryption is disabled.

ZRTTransformEngine is being initialized by default even when DEFAULT_ENCRYPTION=false is set for the account in jitsi-desktop.